requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
requests==2.32.4
python-dotenv==1.0.0
gunicorn==21.2.0